Move MESH_PIPELINE_VIEW_LAYOUT_SAFE_MAX_TEXTURES, MeshPipelineKey
